재고 부족, 품절 ui update함수 분리
